requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.9
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-filter==21.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
flower==1.0.0
humanize==4.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.8.0
Markdown==3.3.7
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.1.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
redis==4.3.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
